requirements.txt
streamlit
google-generativeai
lightweight-charts
streamlit-lightweight-charts